Autocomplete DropDownList based on Chosen js dropDownList
This is DropDownList based on the popular chosen js DropDownList.
It is ASP.NET user control with properites that are the same as in the
javascript version, all chosen documentation can be mapped to this
project.
You can use it freely, just take the ChosenDropDownLiist.dll. include it
in your project and you're good to go.